pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["AgeHT <vladimir.e.pavlov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
pydantic = "^1.10.7"
requests = "^2.30.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
pre-commit = "^3.2.2"
mypy = "^1.3.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
skip-string-normalization = true
skip-magic-trailing-coma = true
exclude = '''
(
  /(
      \.git
    | \.github
    | \.venv
    | deploy
  )/
)
'''
